% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2t.R
\name{d2t}
\alias{d2t}
\title{How to make disease to target}
\usage{
d2t(x, type = "diseaseId")
}
\arguments{
\item{x}{is diseaseId or diseaseName}

\item{type}{the class of x}
}
\value{
target symbol and score
}
\description{
How to make disease to target
}
\examples{
.d2t(x='C0009375',type='diseaseId')
d2t(x=c('C0009375','C0271979'),type='diseaseId')
.d2t(x='Schizophrenia',type='diseaseName')
d2t(x=c('Schizophrenia','beta Thalassemia'),type='diseaseName')
}
