% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2m.R
\name{h2m}
\alias{h2m}
\title{how to make herb to molecule or cid of molecule}
\usage{
h2m(x, type = "latin", output = "cid")
}
\arguments{
\item{x}{an variable}
}
\value{
molecule or cid
}
\description{
how to make herb to molecule or cid of molecule
}
\examples{
.h2m(x='Ziziphi Spinosae Semen',type='latin',output='molecule')
.h2m(x='Ziziphi Spinosae Semen',type='latin',output='cid')
h2m(x=c('Ziziphi Spinosae Semen','Abri Herba'),type='latin',output='molecule')
h2m(x=c('Ziziphi Spinosae Semen','Abri Herba'),type='latin',output='cid')
}
