% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2h.R
\name{m2h}
\alias{m2h}
\title{How to make molecule to herb}
\usage{
m2h(x, type = "cid", OB = "30", DL = "0.18")
}
\arguments{
\item{x}{is molecule or cid}

\item{type}{is the class of x}

\item{OB}{is characteristic value of x}

\item{DL}{is characteristic value of x}
}
\value{
herb information
}
\description{
How to make molecule to herb
}
\examples{
m2h(x=c('thalifendine','MTL'),type='molecule',OB='30',DL='0.18')
m2h(x=c('thalifendine','rosthorin A'),type='molecule',OB='10',DL='0.01')
.m2h(x=c('thalifendine',type='molecule',OB='30',DL='0.18')
m2h(x=c('445858','31423'),type='cid',OB='30',DL='0.18')
m2h(x=c('445858','31423'),type='cid',OB='15',DL='0.02')
}
