% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2m.R
\name{h2m}
\alias{h2m}
\title{how to make herb to molecule or cid of molecule}
\usage{
h2m(x, type = "latin")
}
\arguments{
\item{x}{an variable}
}
\value{
molecule and cid
}
\description{
how to make herb to molecule or cid of molecule
}
\examples{
.h2m(x='Ziziphi Spinosae Semen',type='latin')
.h2m(x='houpu',type='pinyin')
h2m(x=c('Ziziphi Spinosae Semen','Abri Herba'),type='latin')
h2m(x=c('鍘氭湸','榛勮姫'),type='chinese')
}
