% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2t.R
\name{m2t}
\alias{m2t}
\title{how to make molecule to symbol and fullname}
\usage{
m2t(x, type = "cid")
}
\arguments{
\item{x}{is the name or cid of molecule}

\item{type}{is the type of x}
}
\value{
symbol and fullname
}
\description{
how to make molecule to symbol and fullname
}
\examples{
.m2t(x='FER',type='molecule')
.m2t(x='445858',type='cid')
 m2t(x=c('FER','pyrene'),type='molecule')
 m2t(x=c('445858','31423'),type='cid')
}
